requirements.txt
MySQL-python==1.2.5
Naked==0.1.31
Pillow==3.2.0
PyYAML==3.11
SQLAlchemy==1.0.13
amqp==1.4.9
anyjson==0.3.3
argparse==1.2.1
billiard==3.3.0.23
celery==3.1.23
cffi==1.6.0
crypto==1.4.1
cryptography==1.3.2
cycler==0.10.0
enum34==1.1.6
idna==2.1
ipaddress==1.0.16
kombu==3.0.35
matplotlib==1.5.1
numpy==1.11.0
pyasn1==0.1.9
pycparser==2.14
pycrypto==2.6.1
pylib==1.0.11
pyparsing==2.1.4
python-dateutil==2.5.3
pytz==2016.4
requests==2.10.0
shellescape==3.4.1
six==1.10.0
wsgiref==0.1.2